+++ src/main/java/roomescape/dto/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dto;
 
 public class Reservation {
